+++ packages/analytics-plugin-aws-pinpoint/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.20](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.6.19...@analytics/aws-pinpoint@0.6.20) (2021-10-17)
+
+**Note:** Version bump only for package @analytics/aws-pinpoint
+
+
+
+
+
 ## [0.6.19](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.6.18...@analytics/aws-pinpoint@0.6.19) (2021-08-27)
 
 **Note:** Version bump only for package @analytics/aws-pinpoint