+++ packages/analytics-plugin-mixpanel/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/DavidWells/analytics/compare/@analytics/mixpanel@0.3.1...@analytics/mixpanel@0.4.0) (2021-09-25)
+
+
+### Features
+
+* allow mixpanel options to be specified ([a444328](https://github.com/DavidWells/analytics/commit/a444328))
+* enrich mixpanel page events by send all provided properties ([b2ab2bb](https://github.com/DavidWells/analytics/commit/b2ab2bb))
+* support setting of mixpanel page event names ([f7e2fd7](https://github.com/DavidWells/analytics/commit/f7e2fd7))
+
+
+
+
+
 ## [0.3.1](https://github.com/DavidWells/analytics/compare/@analytics/mixpanel@0.3.0...@analytics/mixpanel@0.3.1) (2020-12-23)
 
 **Note:** Version bump only for package @analytics/mixpanel